+++ CHANGELOG.md
@@ -22,6 +22,8 @@ We use *breaking* word for marking changes that are not backward compatible (rel
 ### Changed
 
 - [#2505](https://github.com/thanos.io/thanos/pull/2505) Store: remove obsolete `thanos_store_node_info` metric.
+- [2513](https://github.com/thanos-io/thanos/pull/2513) Tools: Moved `thanos bucket` commands to `thanos tools bucket`, also
+moved `thanos check rules` to `thanos tools rules-check`.
 
 ## [v0.12.1](https://github.com/thanos-io/thanos/releases/tag/v0.12.1) - 2020.04.20